put signers info in root/asfaload.signers/
We put the signers files in a directory, which will siplify the handling
of the multisig as all aggregate signatures are not scoped to a
directory, and we can enforce the fact that a complete aggregate
signature will never be modified. If we create a new signers file, it is
located in a new directory and the existing signatures have no impact.
Previously, for the signers file, we had the the pending signatures
overwriting the current signatures. Now, we can say that a pending
signature file can never overwrite an existing signature file, in both
releases and signers cases.